feat: 리뷰아이템 정렬수정

+++ client/src/Component/DetailPage/StoreHead.jsx
@@ -1,6 +1,8 @@
 import styled from "styled-components";
 import ImgBtn from "../style/ImgBtn";
 import { useState, useEffect } from "react";
+import { useRecoilValue } from "recoil";
+import memberState from "../../state/atoms/SignAtom";
 import { api } from "../../Util/api";
 
 const Container = styled.div`
@@ -42,12 +44,14 @@ const SubInfo = styled.div`
 `;
 
 const StoreHead = () => {
+  // const isLogin = useRecoilValue(isLoginState);
+  const member = useRecoilValue(memberState);
   const [heartIcon, setHeartIcon] = useState(false);
   const [shareIcon, setShareIcon] = useState(false);
 
   const [data, setData] = useState({
     restaurantName: "",
-    tag: [],
+    tagRestaurants: [],
     total_views: "",
     totalFavorite: "",
   });
@@ -56,12 +60,12 @@ const StoreHead = () => {
     const fetchData = async () => {
       try {
         const response = await api.get("/restaurants/1");
-        const { restaurantName, tag, total_views, totalFavorite } =
-          response.data;
+        const { restaurantName, tagRestaurants, total_views, totalFavorite } =
+          response.data.data;
 
         setData({
           restaurantName,
-          tag,
+          tagRestaurants,
           total_views,
           totalFavorite,
         });
@@ -75,20 +79,35 @@ const StoreHead = () => {
 
   const handleHeartIcon = async () => {
     try {
-      if (heartIcon) {
-        await api.delete(`/favorites/1`); //${favorites-id}
-        console.log("즐겨찾기 해제");
-      } else {
+      if (!member.memberId) {
+        console.log("로그인이 필요합니다");
+        alert("로그인이 필요합니다.");
+        return;
+      }
+      if (!heartIcon && member.memberId) {
         await api.post(`/favorites/restaurant/1`); //${restaurant - id}
         console.log("즐겨찾기 저장");
+        setHeartIcon(true);
+      } else {
+        await api.delete(`/favorites/1`); //${favorites-id}
+        console.log("즐겨찾기 해제");
+        setHeartIcon(false);
       }
-      setHeartIcon(!heartIcon);
     } catch (error) {
-      console.error("Error toggling favorite:", error);
+      console.error("즐겨찾기에러:", error);
     }
   };
+
   const handleShareIcon = () => {
     setShareIcon(!shareIcon);
+    if (navigator.clipboard) {
+      navigator.clipboard.writeText(window.location.href);
+      console.log("링크 복사 완료");
+      alert("링크가 복사되었습니다.");
+    } else {
+      console.log("링크 복사 실패");
+      alert("링크 복사에 실패했습니다.");
+    }
   };
 
   return (
@@ -97,9 +116,9 @@ const StoreHead = () => {
         <StoreName>{data.restaurantName}</StoreName>
         <StoreInfo>
           <Tags>
-            {data.tag &&
-              data.tag.map((tag) => (
-                <TagItem key={tag.tagId}>{tag.name}</TagItem>
+            {data.tagRestaurants &&
+              data.tagRestaurants.map((tag) => (
+                <TagItem key={tag.tag.tagId}>{`#${tag.tag.name}`}</TagItem>
               ))}
           </Tags>
           <SubInfo>
